Refactor Table component into modular subcomponents
divisao da table em componentes menores para reutilizacao (e profissionalizacao?)

criacao de um readme explicando
