+++ web/src/components/table/index.js
@@ -1,0 +1,8 @@
+export { default as Table } from '../Table';
+export { default as SearchBar } from './SearchBar';
+export { default as TableHeader } from './TableHeader';
+export { default as TableBody } from './TableBody';
+export { default as TableRow } from './TableRow';
+export { default as TableCell } from './TableCell';
+export { default as Pagination } from './Pagination';
+export { useTable } from './useTable';
